==> Attempt_2/1_Load_merge_files_Train.ipynb <==
#1   ALL clean data is merged into a dataframe known as the training dataset

import numpy as np
import pandas as pd

df_wellfake=pd.read_csv(r"C:\Users\brand\UNISA_Honour_Fake_News_Program\dataset\WELFake_Dataset.csv")
print(df_wellfake.shape)
print(df_wellfake.info())
print(df_wellfake.describe())
print(df_wellfake.columns)


df_wellfake=df_wellfake.drop_duplicates(subset=["title"])

dups=df_wellfake[df_wellfake.duplicated(subset=['text'])]

df_wellfake=df_wellfake.drop(dups.index)

df_wellfake.dropna(axis=0,inplace=True)

print('_______________________________________________________________')
print(df_wellfake.shape)
print(df_wellfake.info())
print(df_wellfake.describe())
print(df_wellfake.columns)

df_wellfake.to_parquet(r"C:\Users\brand\UNISA_Honour_Fake_News_Program\Datasets_Loaded\Well_data.parquet")

df_wellfake.duplicated(subset=['text', 'title']).sum()

df_horn=pd.read_csv(r"C:\Users\brand\UNISA_Honour_Fake_News_Program\dataset\Horn_data.csv")
print(df_horn.isna().sum())
df_horn=df_horn.fillna(0)
print(df_horn.shape)
print(df_horn.info())
print(df_horn.describe())
print(df_horn.columns)
df_horn.to_parquet(r"C:\Users\brand\UNISA_Honour_Fake_News_Program\Datasets_Loaded\Horn_data.parquet")

ukt=pd.read_parquet(r"C:\Users\brand\UNISA_Honour_Fake_News_Program\New_Datasets_For_Final\945z9xkc8d-1\Fake News Dataset\Fake News Dataset\ukt_train.parquet")
ISOT=pd.read_parquet(r"C:\Users\brand\UNISA_Honour_Fake_News_Program\New_Datasets_For_Final\945z9xkc8d-1\Fake News Dataset\Fake News Dataset\ISOT_train.parquet")
jru=pd.read_parquet(r"C:\Users\brand\UNISA_Honour_Fake_News_Program\New_Datasets_For_Final\945z9xkc8d-1\Fake News Dataset\Fake News Dataset\jru_train.parquet")
mc=pd.read_parquet(r"C:\Users\brand\UNISA_Honour_Fake_News_Program\New_Datasets_For_Final\945z9xkc8d-1\Fake News Dataset\Fake News Dataset\mc_train.parquet")
td=pd.read_parquet(r"C:\Users\brand\UNISA_Honour_Fake_News_Program\New_Datasets_For_Final\945z9xkc8d-1\Fake News Dataset\Fake News Dataset\twitter_data.parquet")


td=td.rename(columns={'target':'label'})
td.info()

df_horn_subset = df_horn[[ 'Title', 'Content','Label']].rename(columns={'Title':'title','Content':'text','Label':'label',})


df_wellfake_subset = df_wellfake[[ 'title', 'text', 'label']]


df = pd.concat([ df_horn_subset, df_wellfake_subset,ukt,ISOT,jru,mc,td], ignore_index=True)
df.drop(columns=['title'],inplace=True)

df=df.reset_index(drop=True)




print(df.shape)
print(df.columns)
print(df.info())
print(df.describe())



df.drop_duplicates(subset=['text'],keep='first',inplace=True)
df[df.duplicated(subset=['text'])]

#Merge datasets
df.to_parquet(r"C:\Users\brand\UNISA_Honour_Fake_News_Program\Dataset_Merged_BackUp\Data_merged.parquet")

df.shape
