==> Attempt_2/2_Load_merge_Test.ipynb <==
#2   ALL clean data is merged into a dataframe known as the testing dataset

import numpy as np
import pandas as pd

ukt=pd.read_parquet(r"C:\Users\brand\UNISA_Honour_Fake_News_Program\New_Datasets_For_Final\945z9xkc8d-1\Fake News Dataset\Fake News Dataset\ukt_test.parquet")
ISOT=pd.read_parquet(r"C:\Users\brand\UNISA_Honour_Fake_News_Program\New_Datasets_For_Final\945z9xkc8d-1\Fake News Dataset\Fake News Dataset\ISOT_test.parquet")
jru=pd.read_parquet(r"C:\Users\brand\UNISA_Honour_Fake_News_Program\New_Datasets_For_Final\945z9xkc8d-1\Fake News Dataset\Fake News Dataset\jru_test.parquet")
mc=pd.read_parquet(r"C:\Users\brand\UNISA_Honour_Fake_News_Program\New_Datasets_For_Final\945z9xkc8d-1\Fake News Dataset\Fake News Dataset\mc_test.parquet")


df = pd.concat([ukt,ISOT,jru,mc], ignore_index=True)
df=df.reset_index(drop=True)

df.drop_duplicates(subset=['text'],keep='first',inplace=True)
df[df.duplicated(subset=['text'])]

df.to_parquet(r"C:\Users\brand\UNISA_Honour_Fake_News_Program\Dataset_Merged_BackUp\Data_merged_test.parquet")

df.shape
