==> Attempt_2/3_FeatureCreation.ipynb <==
#3 Create features 

#Possible features
#difference between character removal

#Gaan na die nuwe file kry die test data en run die code op dit ook maak die models dan om beter test scores te kry

import pandas as pd 
import numpy as np

import re
import nltk

from nltk.corpus import stopwords
from nltk.stem import PorterStemmer

nltk.download('punkt')
from nltk.sentiment import SentimentIntensityAnalyzer

from sklearn.feature_extraction.text import CountVectorizer

from nltk.collocations import BigramAssocMeasures, BigramCollocationFinder

from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer
from nltk.tokenize import word_tokenize, sent_tokenize
import textstat


from textblob import TextBlob

import spacy
import en_core_web_sm
nlp = en_core_web_sm.load()

Read in the Training data as well as the test data

df=pd.read_parquet(r"C:\Users\brand\UNISA_Honour_Fake_News_Program\Dataset_Merged_BackUp\Data_merged.parquet")
print(df.info())
df=df.sample(frac=1,random_state=42)#.head(5000)

df_test=pd.read_parquet(r"C:\Users\brand\UNISA_Honour_Fake_News_Program\Dataset_Merged_BackUp\Data_merged_test.parquet")


df_test.info()

df.info()

Word count fuction defined to calculate the word cound for each dataframe

%%time
# Vectorized function to calculate word count

word_count_func = np.vectorize(lambda text: len(word_tokenize(text)))
df['word_count'] = word_count_func(df['text'])


df.head()

%%time
df_test['word_count'] = word_count_func(df_test['text'])


'''
# Vectorized function to calculate average word length
avg_word_length_func = np.vectorize(lambda text: np.mean([len(word) for word in word_tokenize(text)]))
df['avg_word_length'] = avg_word_length_func(df['text'])
'''

Calculate the average sentence length

%%time
# Vectorized function to calculate average sentence length

avg_sentence_length_func = np.vectorize(lambda text: np.mean([len(word_tokenize(sentence)) for sentence in sent_tokenize(text)]))
df['avg_sentence_length'] = avg_sentence_length_func(df['text'])


%%time
df_test['avg_sentence_length'] = avg_sentence_length_func(df_test['text'])


Determine the punctuation and capitalization frequency

%%time
# Vectorized function to calculate the frequency of punctuation marks

punctuation_chars = set('!\"#$%&\'()*+,-./:;<=>?@[\\]^_`{|}~')
df['text_punctuation_frequency'] = df['text'].apply(lambda text: sum(1 for char in text if char in punctuation_chars))


# Vectorized function to calculate the frequency of capitalized words or phrases
vectorized_calculate_capitalization_frequency = np.vectorize(lambda text: sum(1 for word in text.split() if word.isupper()))

df['text_capitalization_frequency'] = vectorized_calculate_capitalization_frequency(df['text'])



df.describe()
df.head(10)

%%time
punctuation_chars = set('!\"#$%&\'()*+,-./:;<=>?@[\\]^_`{|}~')
df_test['text_punctuation_frequency'] = df_test['text'].apply(lambda text: sum(1 for char in text if char in punctuation_chars))
df_test['text_capitalization_frequency'] = vectorized_calculate_capitalization_frequency(df_test['text'])



df.info()

%%time


df['text_before_character_removal']=df['text'].str.len()

def remove_non_characters(text):
    # Remove non-character characters using regular expressions
    cleaned_text = text.str.replace(r'[^a-zA-Z\s]', '', regex=True)
    return cleaned_text

# Apply the function to 'text' column
df['text'] = remove_non_characters(df['text'])

# Convert to lowercase
df['text'] = df['text'].str.lower()

df['text_after_character_removal']=df['text'].str.len()

df.describe()

%%time
df_test['text_before_character_removal']=df_test['text'].str.len()

def remove_non_characters(text):
    # Remove non-character characters using regular expressions
    cleaned_text = text.str.replace(r'[^a-zA-Z\s]', '', regex=True)
    return cleaned_text

# Apply the function to 'text' column
df_test['text'] = remove_non_characters(df_test['text'])

# Convert to lowercase
df_test['text'] = df_test['text'].str.lower()

df_test['text_after_character_removal']=df_test['text'].str.len()

df_test.describe()

%%time
stop_words = set(stopwords.words('english'))
stemmer = PorterStemmer()

def preprocess_text(text):
    # Tokenize the text into individual words
    words = nltk.word_tokenize(text)
    
    # Remove stop words
    words = [word for word in words if word.lower() not in stop_words]
    
    # Perform stemming on each word
    words = [stemmer.stem(word) for word in words]
    
    # Join the processed words back into a single string
    processed_text = ' '.join(words)
    
    return processed_text

df['text_processed'] = ''

def process_row(row):
    text = row['text']
    processed_text = preprocess_text(text)

    
    return pd.Series([ processed_text])

df[['text_processed']] = df.apply(process_row, axis=1)

%%time
stop_words = set(stopwords.words('english'))
stemmer = PorterStemmer()

def preprocess_text(text):
    # Tokenize the text into individual words
    words = nltk.word_tokenize(text)
    
    # Remove stop words
    words = [word for word in words if word.lower() not in stop_words]
    
    # Perform stemming on each word
    words = [stemmer.stem(word) for word in words]
    
    # Join the processed words back into a single string
    processed_text = ' '.join(words)
    
    return processed_text

df_test['text_processed'] = ''

def process_row(row):
    text = row['text']
    processed_text = preprocess_text(text)
    return pd.Series([ processed_text])

df_test[['text_processed']] = df_test.apply(process_row, axis=1)

%%time
calculate_subjectivity_vectorized = np.vectorize(lambda text: TextBlob(text).sentiment.subjectivity, otypes=[float])

# Apply the vectorized function to create a 'subjectivity' column
df['subjectivity_text'] = calculate_subjectivity_vectorized(df['text_processed'])



%%time
# Apply the vectorized function to create a 'subjectivity' column
df_test['subjectivity_text'] = calculate_subjectivity_vectorized(df_test['text_processed'])


%%time
sia = SentimentIntensityAnalyzer()

# Apply sentiment analysis to 'text_processed' column
df['text_sentiment'] = [sia.polarity_scores(text)['compound'] for text in df['text_processed']]

df.head()

%%time
df_test['text_sentiment'] = [sia.polarity_scores(text)['compound'] for text in df_test['text_processed']]

df_test.head()

%%time
def calculate_reading_ease(text):
    if isinstance(text, float):
        return 0  # Return 0 for float values
    # Calculate the reading ease score
    reading_ease = textstat.flesch_reading_ease(text)
    return reading_ease

vectorized_calculate_reading_ease = np.vectorize(calculate_reading_ease, otypes=[float])

# Assuming you have a DataFrame 'df' with 'title_processed' and 'text_processed' columns
df['text_reading_ease'] = vectorized_calculate_reading_ease(df['text_processed'])

df.head()

%%time
df_test['text_reading_ease'] = vectorized_calculate_reading_ease(df_test['text_processed'])

df_test.head()

%%time
def calculate_lexical_diversity(text):
    if isinstance(text, float):
        return 0  # Return 0 for float values
    
    # Tokenize the text into individual words
    words = text.split()
    
    # Calculate the lexical diversity score
    if len(words) > 0:
        lexical_diversity = len(set(words)) / len(words)
    else:
        lexical_diversity = 0
    
    return lexical_diversity

# Define a vectorized version of calculate_lexical_diversity
vectorized_calculate_lexical_diversity = np.vectorize(calculate_lexical_diversity, otypes=[float])

df['text_lexical_diversity'] = vectorized_calculate_lexical_diversity(df['text_processed'])

df.head()

%%time
df_test['text_lexical_diversity'] = vectorized_calculate_lexical_diversity(df_test['text_processed'])

df.head()

df.info()

df_test.info()

%%time
df_backup=df
df_test_backup=df_test
df_backup.info()

df_test_backup.info()

%%time
from sklearn.feature_extraction.text import CountVectorizer

# Initialize CountVectorizer with optional preprocessing options
vectorizer = CountVectorizer(
    stop_words='english',  # Remove common stop words
    max_features=1000,     # Limit the number of features to reduce dimensionality
    binary=True             # Use binary encoding (1 if word is present, 0 otherwise)
)

# Vectorize the text
X = vectorizer.fit_transform(df['text_processed'])

# Create a DataFrame with the BoW representation
bag_of_words_df = pd.DataFrame(X.toarray(), columns=vectorizer.get_feature_names_out())

# Reset the index of df
df.reset_index(drop=True, inplace=True)

# Add a prefix to the BoW columns
new_column_names = ["BOW_" + col for col in bag_of_words_df.columns]
bag_of_words_df.columns = new_column_names

# Save the BoW DataFrame to a Parquet file
bag_of_words_df.to_parquet("BOW.parquet")

# Concatenate df and bag_of_words_df
concatenated_df = pd.concat([df, bag_of_words_df], axis=1)


%%time
X_test = vectorizer.transform(df_test['text_processed'])

# Create a DataFrame with the BoW representation for the test data
bag_of_words_test_df = pd.DataFrame(X_test.toarray(), columns=vectorizer.get_feature_names_out())
df_test.reset_index(drop=True, inplace=True)
# Add a prefix to the BoW columns for the test data
new_column_names_test = ["BOW_" + col for col in bag_of_words_test_df.columns]
bag_of_words_test_df.columns = new_column_names_test

# Concatenate the test_df and bag_of_words_test_df
concatenated_test_df = pd.concat([df_test, bag_of_words_test_df], axis=1)



df_test

concatenated_test_df

#Best of 5 text
bag_of_words_df.shape
bag_of_words_df.head()
#compund words


df.info()


%%time

df.to_parquet(r"C:\Users\brand\UNISA_Honour_Fake_News_Program\Dataset_Merged_BackUp\ForModeling.parquet")

concatenated_df.to_parquet(r"C:\Users\brand\UNISA_Honour_Fake_News_Program\Dataset_Merged_BackUp\ForModeling_BagofWords.parquet")

df.iloc[:, :-1000]

df.to_csv(r"C:\Users\brand\UNISA_Honour_Fake_News_Program\Dataset_Visualisation\Visualize.csv")



df_test.to_parquet(r"C:\Users\brand\UNISA_Honour_Fake_News_Program\Dataset_Merged_BackUp\ForModeling_test.parquet")

concatenated_test_df.to_parquet(r"C:\Users\brand\UNISA_Honour_Fake_News_Program\Dataset_Merged_BackUp\ForModeling_BagofWords_test.parquet")
